requirements.txt
fastapi
uvicorn
python-multipart
pydantic
numpy
pandas
scikit-learn
google
google-genai
python-dotenv
fpdf
pypdf